Completed user jdbc template

See merge request Mikhail_Sboev/Hotel_System!19

+++ src/main/java/com/epam/hotel/daos/RoomDaoTemplateImpl.java
@@ -48,7 +48,7 @@ public class RoomDaoTemplateImpl implements RoomDao {
     }
 
     @Override
-    public boolean delete(int id) {
+    public boolean delete(long id) {
         return false;
     }
 
@@ -63,7 +63,7 @@ public class RoomDaoTemplateImpl implements RoomDao {
     }
 
     @Override
-    public Room getById(int id) {
+    public Room getById(long id) {
         return null;
     }
 }
