+++ src/main/java/com/epam/hotel/services/Service.java
@@ -5,11 +5,11 @@ import java.util.List;
 public interface Service<T> {
     T create(T entity);
 
-    boolean delete(int id);
+    boolean delete(long id);
 
     boolean update(T entity);
 
     List<T> getAll();
 
-    T getById(int id);
+    T getById(long id);
 }
